商品購入機能

+++ app/controllers/items_controller.rb
@@ -23,9 +23,7 @@ class ItemsController < ApplicationController
   end
 
   def edit
-    return if current_user.id == @item.user.id
-
-    redirect_to root_path
+    redirect_to root_path if current_user.id == @item.user.id && @item.purchase.present?
   end
 
   def update
@@ -38,9 +36,7 @@ class ItemsController < ApplicationController
 
   def destroy
     item = Item.find(params[:id])
-      if current_user.id == item.user_id  
-        item.destroy
-      end
+    item.destroy if current_user.id == item.user_id
     redirect_to root_path
   end
